+++ fltk-sys/src/window.rs
@@ -2771,6 +2771,12 @@ extern "C" {
     pub fn Fl_Window_free_position(self_: *mut Fl_Window);
 }
 extern "C" {
+    pub fn Fl_Window_decorated_w(self_: *const Fl_Window) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Window_decorated_h(self_: *const Fl_Window) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Window_find_by_handle(handle: *mut libc::c_void) -> *mut Fl_Window;
 }
 extern "C" {
@@ -3149,6 +3155,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Single_Window_free_position(self_: *mut Fl_Single_Window);
 }
+extern "C" {
+    pub fn Fl_Single_Window_decorated_w(self_: *const Fl_Single_Window) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Single_Window_decorated_h(self_: *const Fl_Single_Window) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Double_Window {
@@ -3509,6 +3521,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Double_Window_free_position(self_: *mut Fl_Double_Window);
 }
+extern "C" {
+    pub fn Fl_Double_Window_decorated_w(self_: *const Fl_Double_Window) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Double_Window_decorated_h(self_: *const Fl_Double_Window) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Menu_Window {
@@ -3863,6 +3881,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Menu_Window_free_position(self_: *mut Fl_Menu_Window);
 }
+extern "C" {
+    pub fn Fl_Menu_Window_decorated_w(self_: *const Fl_Menu_Window) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Menu_Window_decorated_h(self_: *const Fl_Menu_Window) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Gl_Window {
@@ -4206,6 +4230,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Gl_Window_free_position(self_: *mut Fl_Gl_Window);
+}
+extern "C" {
+    pub fn Fl_Gl_Window_decorated_w(self_: *const Fl_Gl_Window) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Gl_Window_decorated_h(self_: *const Fl_Gl_Window) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Gl_Window_flush(self_: *mut Fl_Gl_Window);